requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.2
django-ranged-response==0.2.0
django-registration==3.0
django-simple-captcha==0.5.9
Markdown==3.0.1
Pillow==5.3.0
Pygments==2.2.0
pytz==2018.5
six==1.11.0
